+++ src/utils/cxSdk.js
@@ -37,14 +37,17 @@ export function versionCheck() {
 
 export function cxInit() {
   if (window.location.hostname === 'localhost' || window.location.hostname.includes('ngrok')) {
-    const sdkConf = {
-      baseUrl: 'https://dev-api.cxengagelabs.net/v1/',
-      environment: 'dev',
-      logLevel: 'debug',
-      blastSqsOutput: true,
-      reportingRefreshRate: 10000
-    };
-    CxEngage.initialize(sdkConf);
+    if (window.self === window.top && CxEngage.initialize) {
+      const sdkConf = {
+        environment: 'dev',
+        baseUrl: 'https://dev-api.cxengagelabs.net/v1/',
+        logLevel: 'debug',
+        blastSqsOutput: true,
+        reportingRefreshRate: 10000,
+        supervisorMode: true
+      };
+      CxEngage.initialize(sdkConf);
+    }
   } else {
     loadSdkConf();
     setInterval(versionCheck, 300000);